Create aprilTagWebcam.py
This is the first test of the code it can run on any computer webcam. It can detect angle, depth, and it tells you how far the April-tag is from the center of the webcam. The depth feature only works with 5cm April-tags but you can change that in the code to work with different sizes of April-tags.